+++ nucleotides/command/run_image.py
@@ -1,14 +1,16 @@
 import funcy, os
-import nucleotides.filesystem    as fs
 import biobox.util               as docker
 import biobox.container          as container
 import biobox.cgroup             as cgroup
 import biobox.image.availability as avail
 import biobox.image.execute      as image
-import nucleotides.util          as util
+
+import nucleotides.filesystem          as fs
+import nucleotides.util                as util
+import nucleotides.task.task_interface as interface
 
 def image_type(app):
-    return util.select_task(funcy.get_in(app, ["task", "image", "type"]))
+    return interface.select_task(funcy.get_in(app, ["task", "image", "type"]))()
 
 def image_name(app):
     return funcy.get_in(app, ["task", "image", "name"])
@@ -21,10 +23,6 @@ def image_version(app):
 def image_task(app):
     return funcy.get_in(app, ["task", "image", "task"])
 
-def setup(app):
-    biobox = image_type(app)
-    if hasattr(biobox, 'before_container_hook'):
-        biobox.before_container_hook(app)
 
 def create_container(app):
     avail.get_image(image_version(app))
@@ -57,8 +55,9 @@ def copy_output_files(app):
 
 
 def execute_image(app, docker_timeout = 15, metric_interval = 15, metric_warmup = 2):
-    setup(app)
-    image  = image_version(app)
+    task  = image_type(app)
+    task.before_container_hook(app)
+    image = image_version(app)
 
     app['logger'].info("Creating Docker container from image {}".format(image))
     biobox = create_container(app)